Adaptación base de datos servidor

+++ src/main/java/com/rest/bank/controller/dto/AccountDTO.java
@@ -9,9 +9,9 @@ import java.math.BigDecimal;
 public class AccountDTO {
 
     @NotNull
-    private Long accountNumber;
+    private int accountNumber;
 
     @NotNull
-    private BigDecimal depositAmount;
+    private int depositAmount;
 }
